Worked on deleteCoinAndData function Had to test it tomorrow
